Australian Rain Lab Prediction
Project Scenario
This is Honors Project. Its completion is optional but highly recommended.

Now that you have been equipped with the skills to use different Machine Learning algorithms over the course of five weeks, you will have the opportunity to practice and apply it on a dataset. In this project, you will complete a notebook where you will build a classifier to predict whether there will be rain the following day.

For your project, you will use a rainfall dataset from Australian Government's Bureau of Meteorology , clean the data, and apply different classification algorithm on the data. Alternatively, you can download your data from 
here
. 

You are expected to use the following algorithms to build your models:

1.  Linear Regression
2.  KNN
3.  Decision Trees
4.  Logistic Regression
5.  SVM


The results are reported as the accuracy of each classifier, using the following metrics, when these are applicable:

1. Accuracy Score
2. Jaccard Index
3. F1-Score
4. LogLoss
5. Mean Absolute Error
6. Mean Squared Error
7. R2-Score